+++ .php-cs-fixer.php
@@ -21,6 +21,8 @@ $finder = Finder::create()
     ->in(__DIR__ . '/src/Services/Paysystem/')
     ->in(__DIR__ . '/src/Services/Sale/')
     ->in(__DIR__ . '/src/Services/Task/')
+    ->in(__DIR__ . '/src/Services/Sale/')
+    ->in(__DIR__ . '/src/Services/Calendar/')
     ->name('*.php')
     ->exclude(['vendor', 'storage', 'docker', 'docs']) // Exclude directories
     ->ignoreDotFiles(true)